fbt: split sdk management code
